requirements.txt
-i https://pypi.org/simple
astroid==2.2.5
autopep8==1.4.4
blessings==1.7
bpython==0.18
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
curtsies==0.3.0
defusedxml==0.6.0
django-allauth==0.39.1
django-cors-headers==3.0.2
django-countries==5.3.3
django-filter==2.2.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django==2.2.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djangorestframework==3.9.4
djoser==2.0.3
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
pillow==6.1.0
psycopg2-binary==2.8.4
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pygments==2.4.2
pyjwt==1.7.1
pylint==2.3.1
python-decouple==3.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.1
pyyaml==5.1.2
requests-oauthlib==1.2.0
requests==2.22.0
rest-framework-auth0==0.5.3
six==1.12.0
sqlparse==0.3.0
stripe==2.33.0
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
